add overlay loader and sweetalert2

+++ public/js/script.js
@@ -24,16 +24,25 @@ updateBtn.click(function(e){
         formattedData[item.name] = item.value;
     });
     var url = baseUrl +'products/'+id;
+    $("#update-product-modal").modal('show');
     $.ajax({
         url: url,
         type: 'PUT',
         data:formattedData,
         success: function(msg,statusText,xhr) {
+            $("#update-product-modal").modal('hide');
             getProductAndAddtoTable();
             $("#update-product-modal").modal('hide');
-            $.notify('item update successfully','success')
+            Swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: 'Your work has been updated',
+                showConfirmButton: false,
+                timer: 2500
+              });
         },
         error:function(xhr,text,err) {
+            $("#update-product-modal").modal('hide');
             if(xhr.status && xhr.status==400) {
                 let res = xhr.responseJSON;
                 if(res) {
@@ -66,7 +75,13 @@ updateBtn.click(function(e){
             }
             else {
                 $("#update-product-modal").modal('hide');
-                $.notify('product update fail');
+                Swal.fire({
+                    position: 'center',
+                    icon: 'error',
+                    title: 'update fail',
+                    showConfirmButton: false,
+                    timer: 2500
+                  });
             }
         }
     });
@@ -80,9 +95,11 @@ function addProductHandler(e) {
     formdata.forEach(function(item){
         formattedData[item.name] = item.value;
     });
-    var url = baseUrl + 'products'
+    var url = baseUrl + 'products';
+    $("#add-product-form").LoadingOverlay("show");
     $.post(url, formattedData)
-    .done(function(msg,statusText,xhr){  
+    .done(function(msg,statusText,xhr){ 
+        $("#add-product-form").LoadingOverlay("hide"); 
         if(msg && msg.data) {
 
             productName.val("");
@@ -97,11 +114,18 @@ function addProductHandler(e) {
             totalProduct.removeClass("is-invalid");
             totalProduct.next().text('');
             getProductAndAddtoTable();
-            $.notify('save successfully','success');
+            Swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: 'Your work has been saved',
+                showConfirmButton: false,
+                timer: 2500
+              });
             $("#add-product-modal").modal('hide');
         }
     })
     .fail(function(xhr, status, error) {
+        $("#add-product-form").LoadingOverlay("hide");
         if(xhr.status && xhr.status==400) {
             let res = xhr.responseJSON;
             if(res) {
@@ -132,15 +156,26 @@ function addProductHandler(e) {
                 }
             }
         }
+        else {
+            Swal.fire({
+                position: 'center',
+                icon: 'error',
+                title: 'save unsuccessful',
+                showConfirmButton: false,
+                timer: 2500
+              })
+        }
     });
+
 
 }
 
 function getProductAndAddtoTable() {
     var url = baseUrl + "products";
-
+    $("#product-table").LoadingOverlay("show");
     $.get(url)
     .done(function(msg,statusText,xhr){
+        $("#product-table").LoadingOverlay("hide");
         if(msg && msg.data) {
             var data = msg.data;
             var tbody = $("#tbody");
@@ -152,6 +187,7 @@ function getProductAndAddtoTable() {
         }
     })
     .fail(function(xhr,status,error){
+        $("#product-table").LoadingOverlay("hide");
         console.log(error);
     });
 }
@@ -160,19 +196,57 @@ function getProductAndAddtoTable() {
 $(document).on('click','.delete',function(e){
     let id = $(e.target).data('id');
     var url = baseUrl + "products/"+id;
-    $.ajax({
-        url: url,
-        type: 'DELETE',
-        success: function(msg,statusText,xhr) {
-            console.log(msg,statusText,xhr);
-            getProductAndAddtoTable();
-            $.notify('item delete successfully','success')
+    const swalWithBootstrapButtons = Swal.mixin({
+        customClass: {
+          confirmButton: 'btn btn-success',
+          cancelButton: 'btn btn-danger'
         },
-        error:function(xhr,text,err) {
-            console.log(xhr,text,err);
-            $.notify('product delete fail');
+        buttonsStyling: false
+      })
+      
+      swalWithBootstrapButtons.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'No, cancel!',
+        reverseButtons: true
+      }).then((result) => {
+        if (result.isConfirmed) {
+            $("#product-table").LoadingOverlay("show");
+            $.ajax({
+                url: url,
+                type: 'DELETE',
+                success: function(msg,statusText,xhr) {
+                    $("#product-table").LoadingOverlay("hide");
+                    getProductAndAddtoTable();
+                    swalWithBootstrapButtons.fire(
+                        'Deleted!',
+                        'Your document has been deleted.',
+                        'success'
+                    );
+                },
+                error:function(xhr,text,err) {
+                    $("#product-table").LoadingOverlay("hide");
+                    swalWithBootstrapButtons.fire(
+                        'sorry!',
+                        'delete fail',
+                        'error'
+                      )
+                }
+            });
+        } else if (
+          /* Read more about handling dismissals below */
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          swalWithBootstrapButtons.fire(
+            'Cancelled',
+            'you docuemt is not deleted :)',
+            'error'
+          )
         }
-    });
+      });
 });
 
 //update
@@ -182,20 +256,22 @@ $(document).on('click','.update',function(e){
     let id = $(e.target).data('id');
     var url = baseUrl + "products/"+id;
     $("#update-product-modal").modal('show');
-
+    $("#update-product-modal").LoadingOverlay("hide");
     $.get(url)
     .done(function(msg,statusText,xhr){
+        $("#update-product-modal").LoadingOverlay("hide");
         if(msg && msg.data) {
             $("#display-product-id").text("#"+msg.data._id);
             $('#productName-update').val(msg.data.product_name);
             $('#productUnitPrice-update').val(msg.data.unit_price);
             $('#totalProduct-update').val(msg.data.total_product);
             $('#update-product-form').data('id',msg.data._id);
-            console.log($('#update-product-form').data())
+            console.log($('#update-product-form').data());
         }
     })
     .fail(function(xhr,status,error){
-        $("#update-product-modal").modal('show');
+        $("#add-product-form").LoadingOverlay("hide");
+        $("#update-product-modal").modal('hide');
         $.notify('something went wrong');
     });
 });
